py-src/requirements.txt
glfw
PyOpenGL
numpy
scipy
